move page to correct component folder
