+++ app/renderer/components/page/footer.js
