+++ app/renderer/components/page/info.js
